Updated to latest HttpClient

+++ src/main/java/us/kochlabs/tools/life360/circles/MemberLocation.java
@@ -6,7 +6,7 @@ public class MemberLocation {
 
     public double longitude;
 
-    public int accuracy;
+    public float accuracy;
 
     public int startTimestamp;
